requirements.txt
crewai>=0.28.0
python-dotenv>=0.21.0
pyyaml>=6.0
PyPDF2>=3.0.0
langchain>=0.0.235
langchain-community>=0.0.16
langchain-openai>=0.0.5
openai>=0.28.0
tiktoken>=0.5.0
chromadb>=0.4.18
pydantic>=1.10.8
# PDF orientation detection and correction dependencies
pdf2image>=1.16.3
pytesseract>=0.3.10
pillow>=9.5.00
faiss-cpu>=1.7.4
# API dependencies
fastapi>=0.95.0
uvicorn>=0.21.1
jinja2>=3.1.2
python-multipart>=0.0.6
websockets>=11.0.3
